Fixes in MWindow::fit_autos
Use doubles instead of floats, use update_gui
Change the type of parameters Tracks::get_automation_extents()
and get_extents of various Autos and Automation
Some cosmetics. Use SPDX-License tag.

+++ cinelerra/automation.h
@@ -1,23 +1,7 @@
+// SPDX-License-Identifier: GPL-2.0-or-later
 
-/*
- * CINELERRA
- * Copyright (C) 2008 Adam Williams <broadcast at earthling dot net>
- * 
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation; either version 2 of the License, or
- * (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
- * 
- */
+// This file is a part of Cinelerra-CVE
+// Copyright (C) 2008 Adam Williams <broadcast at earthling dot net>
 
 #ifndef AUTOMATION_H
 #define AUTOMATION_H
@@ -37,16 +21,16 @@
 // Match the clipping at per cinelerra/virtualanode.C which contains:
 //  if(fade_value <= INFINITYGAIN) fade_value = 0;
 //  in reality, this should be matched to a user-defined minimum in the preferences
-#define AUTOMATIONCLAMPS(value, autogrouptype)				\
-	if (autogrouptype == AUTOGROUPTYPE_AUDIO_FADE && value <= INFINITYGAIN) \
-		value = INFINITYGAIN;					\
-	if (autogrouptype == AUTOGROUPTYPE_VIDEO_FADE)			\
-		CLAMP(value, 0, 100);					\
-	if (autogrouptype == AUTOGROUPTYPE_ZOOM && value < 0)		\
+#define AUTOMATIONCLAMPS(value, autogrouptype) \
+	if(autogrouptype == AUTOGROUPTYPE_AUDIO_FADE && value <= INFINITYGAIN) \
+		value = INFINITYGAIN; \
+	if (autogrouptype == AUTOGROUPTYPE_VIDEO_FADE) \
+		CLAMP(value, 0, 100); \
+	if (autogrouptype == AUTOGROUPTYPE_ZOOM && value < 0) \
 		value = 0;
 
-#define AUTOMATIONVIEWCLAMPS(value, autogrouptype)			\
-	if (autogrouptype == AUTOGROUPTYPE_ZOOM && value < 0)		\
+#define AUTOMATIONVIEWCLAMPS(value, autogrouptype) \
+	if (autogrouptype == AUTOGROUPTYPE_ZOOM && value < 0)\
 		value = 0;
 
 struct automation_def
@@ -108,8 +92,8 @@ public:
 		ptstime start,
 		ptstime length,
 		int overwrite = 0);
-	virtual void get_extents(float *min, 
-		float *max,
+	virtual void get_extents(double *min,
+		double *max,
 		int *coords_undefined,
 		ptstime start,
 		ptstime end,
